+++ task.go
@@ -330,7 +330,7 @@ func (e *Executor) runDeferred(t *ast.Task, call *ast.Call, i int, deferredExitC
 	}
 
 	cmd := t.Cmds[i]
-	vars, _ := e.Compiler.FastGetVariables(origTask, call)
+	vars, _ := e.Compiler.GetVariables(origTask, call)
 	cache := &templater.Cache{Vars: vars}
 	extra := map[string]any{}