+++ variables.go
@@ -309,6 +309,10 @@ func itemsFromFor(
 					} else {
 						values = asAnySlice(strings.Fields(value))
 					}
+				case []string:
+					values = asAnySlice(value)
+				case []int:
+					values = asAnySlice(value)
 				case []any:
 					values = value
 				case map[string]any: